Implemented handleReportPriceMessage() by intention. Need to interpret the leading bidder for a "report price" message.
I'm uncomfortable with this method signature:

    public static String leadingBidderAccordingTo(Message reportPriceMessage);

In particular, the name of the variable is more specific than the name of the class, meaning I have to either check type or assume type -- a Morton's fork. I'd rather unpack the message, convert it to a meaningful object, then ask that to do work. Factory, maybe, of Message -> implementation of AuctionAction? Something like that.